AI_Mail_Assistant/requirements.txt
# Gmail/Email API
google-api-python-client==2.107.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
sqlite3-binary==0.0.1

# LLM Integration
openai==1.3.7
langchain==0.0.344
langchain-openai==0.0.2

# Tool Integrations
slack-sdk==3.23.0
google-api-python-client==2.107.0  # For Google Calendar API
requests==2.31.0
beautifulsoup4==4.12.2

# Utilities
python-dotenv==1.0.0
pydantic==2.4.2
fastapi==0.104.1
uvicorn==0.24.0
schedule==1.2.1
